fixed an issue with system send Email having a faulty sender stored in the db maintenance rate and poola re now displayed as ints in the structures status window and not as floats with decimal point

+++ src/ZoneServer/UIManager.cpp
@@ -685,11 +685,11 @@ void UIManager::createNewStructureStatusBox(UICallback* callback,PlayerObject* p
 
 	minutes = (uint32)(maint/(rate/60));
 	
-	sprintf(text,"Maintenance Pool: %f(%u days, %u hours, %u minutes)",(float)structure->getCurrentMaintenance(),days,hours,minutes);
+	sprintf(text,"Maintenance Pool: %u(%u days, %u hours, %u minutes)",(uint32)structure->getCurrentMaintenance(),days,hours,minutes);
 	attributesMenu.push_back(text);
 
 	//Maintenance rate
-	sprintf(text,"Maintenance Rate: %f/hr",rate);
+	sprintf(text,"Maintenance Rate: %u/hr",(uint32)rate);
 	attributesMenu.push_back(text);
 
 	//Power Pool
